Added feature to generate and use a golden image
This commit adds the posibility to generate and use golden images.

The major benefit of golden images is that it already has done
time-consuming tasks (such as installing updates and cleaning updates
files). The images are in VHDX format. Warning: The golden images are
not deployment ready because they aren't sysprepped.

In order the use them, the New-WindowsFromGoldenImage function must be
used. More details can be found in the Examples and function synopsys.

Signed-off-by: Costin Galan <cgalan@cloudbasesolutions.com>