Added the possibility to generate golden images
This commit adds the option to generate golden images, which are
cloud images where the long operations are already done (installing
updates and cleaning the extra files).

Note: This image is not ready to be used in a cloud. This image can
be used as a starting point to generate a fully working image.

Signed-off-by: Costin Galan <cgalan@cloudbasesolutions.com>