Add config to ignore extra_packages errors
When installing a directory full of updates, such as from wsusoffline,
there are inevitably some packages which are "not applicable to this
image". When this happens, DISM returns a nonzero exit code and the
image build fails. This configuration allows the user to ignore these
errors and proceed with the image build anyway.