licensing: added option to use kms keys
To fully automate the Windows image creation, for non-core
Windows Server images the KMS key needs to be set in the
Unattend.xml so that no manual action is required to
press on 'Skip License' button.

Currently, only Server Standard and Datacenter (core) are enabled.

To enable this option, the value of the config option
product_key needs to be set to default_kms_key.