Reset BCD store for images created from gold
If multiple images are created at the same time from a gold VHD,
there is a chance that some of them do not boot because HyperV
boot environment does not recognize the boot device anymore
with error code 0xc00000e.

Resetting the BCD store to locate the os device and resetting the
VHD GUID should solve the issue.