Added option to install Qemu Guest Agents
Added option to install Qemu Guest Agents:
[custom]
install_qemu_ga=False

If set to True, it will download and install the Qemu Guest
Agents from Fedora website. The base path for the agents is
https://fedorapeople.org/groups/virt/virtio-win/direct-downloads/archive-qemu-ga/qemu-ga-win-100.0.0.0-3.el7ev/
The downloaded MSI will be chosen depending on the architecture of the
image, x86 or x64.

For example, on x64 images, the following MSI installer will be
downloaded and used:
https://fedorapeople.org/groups/virt/virtio-win/direct-downloads/archive-qemu-ga/qemu-ga-win-100.0.0.0-3.el7ev/qemu-ga-x64.msi
By default, the imaging tools will not install the Qemu Guest
Agents.

The MSI will install two services on the Windows image:

 * QEMU-GA
 * QEMU Guest Agent VSS Provider

The Qemu Guest Agents are useful if the Windows instance runs
on OpenStack (KVM hypervisor is required).

If the Windows instance is managed by OpenStack, runs on KVM and
QEMU-GA service is running, the user password can be changed using
the OpenStack Nova command:

  nova set-password <instance-id>

Requirement1: if the nova command fails, the instance might enter an error
state and you need to reset the instance state:

  nova reset-state <instance-id> --active

Requirement: the Glance image needs to modified in the following way,
BEFORE the instance gets created:

  glance image-update <image_id> --property hw_qemu_guest_agent=True \
  --property os_admin_user=Admin --property os_type=windows

`os_admin_user` is the user that will have the password changed.
Defaults to user 'root'. If 'os_type' property is set
to  'windows', the default user is 'Administrator'.
If you modify the required Glance image properties after you created
the instance, the nova set-password will not function.