Add option to set cloudbase-init service to auto-delayed start
Added option to set cloudbase-init service to auto-delayed start.

```ini
[cloudbase_init]
cloudbase_init_delayed_start=False
```
Sometimes we need cloudbase-init service to be set to auto delayed
start, to make sure that all the Windows services are up and running
when cloudbase-init runs. This scenario might fix the problems with
NETLBFO bond, when cloudbase-init tries to create the bonds, but the
bonding networking services are not ready and silent failures can occur.

Fixes:
https://github.com/cloudbase/windows-openstack-imaging-tools/issues/341